Rename the podcast gallery page file
