+++ src/lib/pages/podcast-gallery/index.ts
@@ -1,2 +1,2 @@
-export { default as PodcastGalleryPage } from './ui/index.svelte';
+export { default as PodcastGalleryPage } from './ui/podcast-gallery-page.svelte';
 export { loadPodcasts } from './api/load-podcasts';
